pub_sub example
