+++ examples/pub_sub/README.md
@@ -1,3 +1,15 @@
-## Sender receiver example
- Basic example that show how components that are not aware of each other can communicate
- using an agent.
+# Pub Sub Example
+
+This is currently a technical demonstration of agents.
+
+## Concepts
+
+The example has two components, which communicate through a "broker" agent
+as opposed to the traditional method using component links.
+
+## Improvements
+
+As it stands, this example uses a great amount of code to do very little.
+The concept should be applied to a more elaborate use case.
+
+This could also be merged into the [nested_list](../nested_list) example to remove the need for `WeakComponentLink`.
