+++ examples/pub_sub/index.html
@@ -1,0 +1,9 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="utf-8" />
+    <title>Yew • Pub Sub</title>
+  </head>
+
+  <body></body>
+</html>
