+++ examples/pub_sub/src/subscriber.rs
@@ -15,24 +15,24 @@ impl Component for Subscriber {
     type Message = Msg;
     type Properties = ();
 
-    fn create(_: Self::Properties, link: ComponentLink<Self>) -> Self {
-        let callback = link.callback(Msg::NewMessage);
-        let _producer = EventBus::bridge(callback);
-        Subscriber {
-            message: "No message yet.".to_string(),
-            _producer,
+    fn create(_props: Self::Properties, link: ComponentLink<Self>) -> Self {
+        Self {
+            message: "No message yet.".to_owned(),
+            _producer: EventBus::bridge(link.callback(Msg::NewMessage)),
         }
     }
 
-    fn change(&mut self, _: Self::Properties) -> bool {
-        false
+    fn change(&mut self, _props: Self::Properties) -> ShouldRender {
+        unimplemented!()
     }
 
     fn update(&mut self, msg: Self::Message) -> ShouldRender {
         match msg {
-            Msg::NewMessage(s) => self.message = s,
+            Msg::NewMessage(s) => {
+                self.message = s;
+                true
+            }
         }
-        true
     }
 
     fn view(&self) -> Html {
